MAINT refactor stuff from base.py into utils.py
