+++ mne/io/ctf/ctf.py
@@ -13,7 +13,8 @@ import numpy as np
 from ...utils import verbose, logger
 from ...externals.six import string_types
 
-from ..base import _BaseRaw, _mult_cal_one, _blk_read_lims
+from ..base import _BaseRaw
+from ..utils import _mult_cal_one, _blk_read_lims
 
 from .res4 import _read_res4, _make_ctf_name
 from .hc import _read_hc
