+++ mne/io/kit/kit.py
@@ -21,7 +21,8 @@ from ...coreg import fit_matched_points, _decimate_points
 from ...utils import verbose, logger
 from ...transforms import (apply_trans, als_ras_trans, als_ras_trans_mm,
                            get_ras_to_neuromag_trans, Transform)
-from ..base import _BaseRaw, _mult_cal_one
+from ..base import _BaseRaw
+from ..utils import _mult_cal_one
 from ...epochs import _BaseEpochs
 from ..constants import FIFF
 from ..meas_info import _empty_info, _read_dig_points, _make_dig_points
